컨트롤러 분리: TouristController와 FoodController로 분리
